requirements.txt
allure-behave
behave
selenium
webdriver-manager
python-dotenv